// src/app/interfaces/hotel.interface.ts
export interface Hotel {
  id?: number;
  nombre: string;
  direccion: string;
  telefono: string;
  categoria: number;
  estado: boolean;
}

// export interface CreateHotelRequest {
//   nombre: string;
//   direccion: string;
//   telefono: string;
//   categoria: number;
//   estado: boolean;
// }